VpcFrugal/requirements.txt
aws-cdk.core
aws-cdk.aws-lambda
aws_cdk.aws_lambda_python
aws_cdk.aws_iam
aws_cdk.aws_events
aws_cdk.aws_events_targets
aws_cdk.aws_logs
-e .
